wc - Reestructured enpoint into input and result
